backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
twilio==9.0.4
httpx==0.27.0
apscheduler==3.10.4
